cleanup(storage)!: uploads track committed_size
Fixing how the `storage` client library deals with reponses lacking a
`Range` header will break any existing unit tests that mock a
`ResumableUploadSession`. This creates an opportunity to cleanup the
semantics of `ResumableUploadResponse`. This change performs that
cleanup.

**BREAKING CHANGE:** with this PR any use of the
`storage::internal::ResumableUploadResponse` type require changes.
Applications should have little need for this type, outside mocks, so
the changes should not affect production code.

Nevertheless, we apologize for the inconvenience, and while we would have
preferred to avoid breaking changes, it was inevitable to introduce some
breaking changes to fix a data loss bug.

If you are affected by this change, we expect that any existing tests
will fail to compile. This was an intentional choice, the semantics of
some values are changing, we think it is better to have these issues
detected at compile-time rather than during test execution.

We have updated the examples to use the new mocks, and the release notes
include more details on how to change any affected tests.