fix(storage)!: handle missing range header in uploads (#7877)
Resumable upload responses use the `Range` header to report how many
bytes were committed. Sometimes, maybe once every few million uploads,
the header is missing. This indicates that nothing was committed.
The library was treating this as a permanent error. The correct behavior
is to query the status of the upload and try again.