feat(generator): streaming-write RPCs for Stubs (#7941)
Change the generator to support streaming write (aka client-side
streaming) RPCs in the `*Stub` class and its decorators. Changed the
golden files to include one such method. Change the unit tests for the
golden files to test the generated code.

The generated libraries are unaffected by this change, as none of the
libraries we have uses streaming write RPCs.