cleanup(generator): separator before code
In some places we generated code and then added a separator line in
anticipation of the next code fragment.  That leaves unsightly blank
lines where not needed. In other places we inserted the separator before
the code, which is easier to reason about ("I am about to insert code, I
should add a separator"), and the unneeded lines before a declaration
are not that unsightly, maybe?