ci: stop cmake-install build on the first error (#7996)
We often run this build manually. It was tedious to manually check the
exit code, so now the script prints "SUCCESS" when it completes
successfully. It was also tedious to search the log for errors, so now
it stops on the first failure, like most of our CI scripts do.