feat(generator): handle proto fields named with C++ keywords
Motivated by a new service with a proto containing a "namespace"
field.  The C++ proto compiler adds a trailing underscore to such
fields and their associated accessors, so we should do the same.
And we should also do it in the parameters to "method_signature"
overloads, and in generated @param annotations.

If anyone has any better ideas for working around the cmake issue,
or for some alternative method to access libprotoc's cpp keyword
mapping, I'm all ears.