feat(generator): handle proto fields named with C++ keywords (#8063)
Motivated by a new service with a proto containing a "namespace"
field. The C++ proto compiler adds a trailing underscore to such
fields and their associated accessors, so we should do the same.
And we should also do it in the parameters to "method_signature"
overloads, and in generated @param annotations.