test(spanner): speed up test with emulator (#8072)
The `rpc_failure_threshold_integration_test` was running very slowly
against the emulator.  The test is not that useful against the emulator,
but we do not want it to rot.  I changed the test to detect that it is
running against the emulator, and then simply reduce the running time, and
ignore the statistical results as they become unreliable.