cleanup(generator): simplify golden files upkeep (#8067)
Change the generator so we can use as single command to update the
golden files.  With this in place, comparing the golden files becomes
very similar to the process of comparing the generated files, so I
change the builds to do that instead.  That produces nice diffs when the
golden files change, as opposed to the unreadable output from
googletest.  I then removed the integration test program for the
generator, since it was no longer needed.

I also documented how to manually run the generator to produce the
golden files. Before, the only known way to do it involved running one
of the builds in "--docker-shell" mode, waiting for a full build, for
the unit tests to pass (note this required the golden files to compile),
then for the integration test to fail, then copy the golden files from
`/tmp`.  During development the golden files would (often) not compile,
so one needed to revert the golden files and start over.