fix(generator): fix options handling in SetIamPolicy() OCC loop
Fix `Options` handling in the generated OCC-loop `SetIamPolicy()`
overloads.  We were failing to pass the operation options through
to the sub-`GetIamPolicy()`/`SetIamPolicy()` calls.  Continuing to
call back into the client layer from the OCC loop is suboptimal
however, as we would need to make a copy of the `Options` only to
create new `OptionsSpan`s with the same values.  So, call through
to the connection layer directly, avoiding the need to deal with
`Options` at all.

Then add tests, for each variety of generated RPC, to check that
options are correctly propagated and overridden through connection,
client, and operation layers.