chore: prefer @platforms rather than @bazel_tools
The former is the preferred way to specify platforms. Note that in the
testing_util case, I dropped the support for freebsd and openbsd. I'm
not sure why those cases were there or if we need them. We don't seem to
have those cases elsewhere. I can add them back if they are needed.