ci: reduce CTest output on interactive builds
This will make it easier to find stuff when running the builds locally.
I expect the CI builds are not affected, and still have long outputs. I
had to refactor some stuff in the CI scripts to make sure ctest got the
same args everywhere, that required favoring `-j` over `--parallel`
because `--parallel` is not supported by the oldest version of CTest we
happen to use.

I also found that gcc-54.sh was still around, but unused, easier to
remove it than fixing it.