chore: upgrade and fix typos-cli
This upgrades us to the newest version of typos-cli. The previous
version we were using (1.3.1) was also broken, but we didn't notice
becuase I think all of our `gcb-checkers` docker images had an old
version that worked. Aside: I think somehow the cached docker images
don't get invalidated properly.

Anyway, the string "ba" showed up as a typo (in the old and new
version). I changed it to "bx", which I think makes the word weird
enough that `typos` doesn't flag it as a typo. I also upgrade to a newer
version of typos to make sure all the docker images get rebuilt.