feat(generator): support parameters named "options"
The googleapis RPCs can define "method signatures" which decompose a
proto `request` object into its key parameters. The generator maps these
to overloads, and adds the `Options options` parameter. This does not
work when one of the arguments defined in the method signature is
already called `options`.  This changes the extract `Options options`
parameter to `Options opts`, which, as of this writing, does not have
any conflicts in the googleapis protos.