feat(generator): support `x-goog-user-project` (#8245)
Many services cannot be used directly from a user account. User accounts
have no default project associated with them for quota and billing, and
some services do not have any resources attached to a specific project
either. For example, compute-only services like Text-to-Speech. In this
case the recommendation is to use the x-goog-user-project header, or use
service account impersonation. The latter is fairly difficult, and would
make the "quickstart" not very quick.

This PR introduces a new option to configure this header. The default
comes from an environment variable.