refactor: use helpers to populate default options
All the libraries need to initialize some common values in the default
`Options`. For example, `EndpointOption`, or `TracingOptions`. With this
change all the generated libraries use some common helpers to do this
work.

Things are a bit complicated because the common options are split across
the `google_cloud_cpp_common` and `google_cloud_cpp_grpc_utils`. I wrote
two separate functions to make that doable.