doc: update stale warnings for macOS
We have not needed the `OPENSSL_ROOT_DIR` environment variable set for a
long time, since we stopped doing super builds of gRPC at least.

Turn the message to a `WARNING`, as ee do recommend setting the CMake
variable of the same name on Apple, because the default OpenSSL is not
functional.

Let CMake format the message because it does a better job: it uses the
terminal width to break lines.