test: enable oslogin and shell quickstarts in CI (#8371)
I think the best we can do for these services is run them, and verify we
get some kind of error.

These services are intended for *users* to retrieve their preferences
from GCP.  They cannot be used from service accounts, and only the user
can access their preferences.  That all sounds fine. Our integration
tests authenticate with a service account (as they should), and thus
cannot retrieve their own preferences (they are not a user). Even if we
created a demo or fake user, the service accounts cannot access the
preferences for such users, and I do not think we would want to do that
anyway.