ci: exclude any googleapis.patch from whitespace fixing (#8374)
In a preview branch we may have merged a `googleapis.patch` file
to support a new feature that uses unreleased protos.  This patch
file often has lines that contain a single space because there are
empty context lines in the diff.  We do not want to trim those
trailing spaces, so exclude `googleapis.patch` from the process.