fix: restore Options over deletion of StreamRange<T>::reader_ (#8403)
Restore the creation-time `Options` in the `StreamRange<T>` destructor
while explicitly clearing the `reader_` functor.  If the reader has not
reached the end of the range, then deleting it may send a best-effort,
out-of-band cancel on the call, and we should restore the options when
re-entering the library.  Add a test for the destructor case.  This is
a followup to #8256.

Also take this opportunity to delete some unused type aliases I stumbled
upon during this work.