feat: better support for PSC and VPC-SC
When using [PSC] (Private Service Connect) or [VPC-SC] (Virtual Private
Cloud - Service Controls) we need to specify the `authority` in each
gRPC request. This is the HTTP/2 analog of a `Host` header, and allows
applications to access a service while using a custom endpoint. For
example, the target endpoint could be `pubsub-custom.p.googleapis.com`,
but the authority field remains `pubsub.googleapis.com`.

[PSC]: https://cloud.google.com/vpc/docs/private-service-connect
[VPC-SC]: https://cloud.google.com/vpc-service-controls