feat(pubsub): support x-goog-user-project (#8456)
Change the metadata decorators to inject `x-goog-user-project` from the
current option span.

For the admin APIs, use the per-call options to configure a span in the
`*Client` class.  For `pubsub::Publisher` and `pubsub::Subscriber` it is
less clear which call "triggers" an RPC, and far less clear how to merge
options from multiple RPCs.  I have decided to use the per-connection
options in this case. We can always add more fine-grained options later,
but this takes care of the common case.