feat(storage): support AuthorityOption
If present, use the `AuthorityOption` to configure the `Host: ` header
or the `.set_authority()` attribute in `grpc::ClientContext`. By
default, configure `AuthorityOption` to be `storage.googleapis.com`,
applications can override this when initializing the `storage::Client`.

There is already an issue open to add per-call `Options` that would
provide more fine-grained control of this field.