ci: automatically update GA libraries in top-level README
Use the top-level `BUILD.bazel` as the source of truth for GA libraries.
That requires installing `bazelisk` in the `checkers` build image.

Once that is working, we can automatically update the section of the
README describing which libraries are GA: just iterate over the list,
extract the service name / description from each library README, and
then print out the right links. Note that I discovered incorrect links
in the process.