impl: a benchmark and two small (non-)optimizations for Options (#8487)
With the growing ubiquity of `Options`, it seems like a good
idea to keep an eye on its performance.  To that end, add a
micro benchmark which can be expanded as we see fit.  For now,
simply benchmark `get<T>()` performance on both present and
default option types.

Time:
The original motivation for this PR was removing the function-
static synchronization from the "present" path, however the
benchmark indicates that this is only a slight win (<0.5%),
and it seems to cost us a similar amount in the "default"
path.  At least it makes those cases similar in performance,
which is a good thing given that it is unclear which is more
common.  (I guess it is also a good thing to confirm that
function-static synchronization is fast.)

Space:
Template the allocation of the default value for an option on
`FooOption::Type` rather than `FooOption`.  This means, for
example, that we only allocate a single `std::string` default
value for all `std::string`-based options.

Also add a note about how `unordered_map<std::type_index, ...>`
works when indexed by `typeid(T)`.