fix(spanner): capture the ReadRequest by value in the reader factory
The `ReadRequest` captured within `PartialResultSetReaderFactory`
needs to be valid after `ConnectionImpl::ReadImpl()` returns because
`PartialResultSetResume::Read()` will use the factory/request, along
with an updated `resume_token`, to create a new reader when the
previous reader finishes with a retry-able status.  So, the request
needs to be captured "by value".

However, `ConnectionImpl::ReadImpl()` also needs to be able to
update the transaction field within that `ReadRequest` (with the
result of an implicit `BeginTransaction()`) should the original
request fail or otherwise not yield metadata.  So, we also need to
hold a reference to that captured value.

To satisfy both requirements, heap allocate the `ReadRequest` and
share its ownership between the factory and `ReadImpl()`.

To test this, we modify `ConnectionImplTest.ReadSuccess` to use a
`PartialResultSetReader` that only delivers partial results before
failing with a retry-able status, causing `PartialResultSetResume`
to re-invoke the factory long after `ConnectionImpl::ReadImpl()`
has returned.  Without the fix, this modification would elicit a
"use-of-uninitialized-value" error from the MemorySanitizer.

This problem has been with us since #4722 (that I reviewed), which
fixed the "needs to be able to update the transaction" part by
capturing the stack-allocated `ReadRequest` by reference.  I guess
that means the factory is rarely re-invoked in practice.