test(storage): stream benchmark results
I am starting to run benchmarks that take hours to complete.  Sometimes
it is desirable to capture partial results from these benchmarks to
debug the analysis scripts, or even run preliminary analysis. With this
change the throughput-vs-cpu benchmark streams its results as it runs,
instead of printing the results at the end of the test.