ci: wait longer for the pubsub emulator to start listening
Loop 60 times (up from 12) while waiting to see the "listening on"
message from the pubsub emulator log.

For both pubsub and spanner emit the "Cannot determine port" error
message with a timestamp, so that it is easy to see just how long
we waited.

Finally, for spanner, dump the emulator log when we fail to find a
port, so that we might have some idea of what went wrong.