fix(storage): return an error on short writes (#8562)
Very rarely the retry policy is exhausted during an upload. If this
happens when the last request was a query to find the status of the
upload the `UploadChunk()` succeeds with a short write.

We need to determine if that is a good idea (it might be). In the
interim, we need to signal to the application that the stream is no
longer usable. They can use `.Suspend()` to reset the stream (maybe in
another process) if that can be useful.