feat(spanner): integration tests for Database.database_dialect
Extend existing tests to also cover `database_dialect`by setting it
in the CreateDatabaseRequest, and expecting it in Database and Backup
responses.  Include allowances for these being invisible to the Spanner
emulator, and for pending bug #8573.

In particular, BackupIntegrationTest.BackupRestore leaves the dialect
unspecified in CreateDatabase() (but expects to see GOOGLE_STANDARD_SQL
in responses), while it is set explicitly to GOOGLE_STANDARD_SQL in
DatabaseAdminClientTest.VersionRetentionPeriodCreate, and to POSTGRESQL
in BackupExtraIntegrationTest.BackupRestoreWithCMEK.

DatabaseAdminClientTest.DatabasePostgreSQLBasics is a new test case
that also exercises the non-backup parts of an explicit POSTGRESQL
setting.

[Aside: These are only in the tests for the generated database admin
client.]

Finally, add ClientIntegrationTest.DatabaseDialect to enumerate the
`database_dialect` value from `INFORMATION_SCHEMA.DATABASE_OPTIONS`
for a default-dialect database.