test(spanner): change CopyBackup() sample tag
The C++ Cloud Spanner client was the one that used the specified tag
for the CopyBackup() sample.  Someone else got it wrong originally it
seems, and then the remaining languages just copied that.

The easiest path forward was to change the specification, so this now
brings the C++ sample into line with the new reality.