ci(spanner): decrease LRO polling-time limit for backup tests
Decrease the polling-time limit from 3 hours to 90 minutes.  The
test infrastructure has a 3h timeout itself, and if that hits first
then we won't see the RPC tracing needed to figure out what is going
wrong.  Successful runs indicate that the backup should only take
about 30-45m, so the 90m limit should still be sufficient to only
catch problematic cases.

Part of #8616 and #8628.