ci(spanner): tweak response to polling timeout in CreateBackup() test (#8674)
Reacting to a "terminated by polling policy" error is difficult at
best as the caller won't know the actual state of the operation, and
will have little they can do to regain state.  In the case at hand,
we stop waiting for the CreateBackup(), but it continues none the
less, and we no longer have a handle by which the operation can be
canceled.  Attempting to drop the database while a backup operation
is pending will fail, so don't even try.  Instead, we rely upon the
short expiration time of test backups, and the eventual cleanup of
test databases (2 days) to reclaim the lost resources.

In general, the client library probably shouldn't be synthesizing
a return status for an operation.  Rather, it should be accurately
reflecting the state of the service. Leave a comment to that effect,
which we can discuss and follow up on later.

Also, the previous increase in the "lastN" tracing to 256 was not
adequate to capture a ~100x polling loop, so bump it to 1024. There
are 6 trace events per loop, so that should be sufficient.