test(spanner): revert to using hyphen separators in database names (#8685)
Now that I have abandoned trying to parameterize tests/samples on
database dialect, it is straightforward to choose the right quoting
convention needed to include hyphens in database names in the CREATE
DDL statement.  So, resume using "db-YYYY-MM-DD-" prefixes for random
database names---they are easier to produce and read.  This reverts
most of #8547.

Also change `RandomDatabasePrefixRegex()` to return a regex that
indeed only matches the random-database prefix, and which therefore
should be passed to `std::regex_search()`.