fix(common): revamp the async polling loop part deux
Do not fabricate a deadline-exceeded `Status` when the polling policy
says enough is enough.  That left the operation in an unknowable state.
Rather, cancel the operation and wait for the next poll to yield us an
accurate status (which may end up being OK).

That is, a polling limit is really a request for auto-cancellation.

(See #7762 for the part-un work on ensuring an explicit cancel yielded an
accurate status.)

In the case of the Spanner backup tests, for example, this means we can do
something meaningful after hitting the poll limit of a `CreateBackup()`
call, as we'll know that the backup is no longer pending.

Update the poll-loop mocking tests to account for the new sequence of RPCs.