feat(spanner): add spanner::Value support for Type.type_annotation
Add support for passing arguments through the `absl::optional` and
`std::vector` constructors for `spanner::Value`, to be applied when
building the prototypical `T` value used to set the protobuf `Type`
member should the `optional` or the `vector` be empty.

The `optional` case means we also need to pass `T` construction
arguments through `MakeNullValue<T>()`.

The Spanner C++ types are in one-to-one correspondence with the
protobuf `Type.code`.  But `message Type` now has a new field,
`type_annotation`, that disambiguates the SQL type that Spanner
will use for particular values of the C++ type.  Therefore, we need
to be able to set the `type_annotation` field when originating such
values.

Normally this isn't a problem as we have the value at hand, but in
the cases of an empty `optional` or `vector` we need to construct
a `T`.  Previously we used the default constructor to do that, but
now we can pass through construction arguments, which can influence
the chosen `type_annotation`.

For example, `spanner::MakeNullValue<T>(args...)`.

So, legitimize the use of the `MakeTypeProto()` argument value
when building the `message Type`.  Previously we only used its C++
type for overload dispatch.

Also clarify that `type_annotation` does not affect the result of
`TypeProtoIs()`, which only concerns the C++ type.

The first use of these construction arguments will be in an upcoming
change to `spanner::Numeric`, but for now we just test that arguments
can be passed in the right places.