feat(storage): new option to disable decompressive transcoding
For objects stored in gzip format (and with contentEncoding == "gzip")
GCS over HTTP automatically decompresses the object during download.
Some applications may want to read the object in compressed format.
Support this use-case with a new option (`AcceptEncoding`) for
`Client::ReadObject()`, and a new helper function (`AcceptEncodingGzip()
`) that returns this option with the correct value.