feat(spanner): add spanner::PgNumeric with larger range/resolution
The Google Spanner variety of NUMERIC has 29 decimal digits of integer
precision and 9 decimal digits of fractional precision, but the PostgreSQL
variety has 131072 and 16383 digits respectively (and supports NaN).

So:
 - Introduce `spanner::Decimal<DecimalMode>` which is templated to
   handle both varieties,
 - Re-brand `spanner::Numeric` as an alias for the Spanner variety.
 - Add `spanner::PgNumeric` as an alias for the PostgreSQL variety.

This pivots from a previous decision to handle both varieties is a single
C++ type.  Now they are unrelated, and can be used to set both `Type.code`
and `Type.type_annotation` in an upcoming change to `spanner::Value`.