feat(spanner): add support for spanner::PgNumeric
The Google-Spanner variety of NUMERIC has 29 decimal digits of integer
precision and 9 decimal digits of fractional precision, while the new
PostgreSQL variety has 131072 and 16383 digits respectively (and supports
NaN).

So:
 - Introduce `spanner::Decimal<DecimalMode>`, which is a templated
   refactoring of the old `spanner::Numeric` code but that handles
   either precision specification.
 - Re-brand `spanner::Numeric` as an alias for the Google variety of
   `Decimal`.
 - Add `spanner::PgNumeric` as an alias for the PostgreSQL variety.

This pivots from a previous decision to handle both varieties is a single
C++ type.  Now there are two unrelated types, which will be used to set
both the `Type.code` and `Type.type_annotation` protobuf fields in an
upcoming change to `spanner::Value`.