doc(pubsub): deprecate `SubscriberOptions`
As usual, we cannot mark the class itself as deprecated, because it is
used in functions needed for backwards compatibility (even though those
functions are deprecated too). What we can do is mark all member
functions, including constructors, as deprecated.  We also document the
class and all its member functions as deprecated too.