refactor(spanner): use struct for transaction callback parameters
Introduce `TransactionContext` to encapsulate the read-only callback
parameters to the transaction visitation functor.

When we added the first parameter, `seqno`, we said, "Great, a single
extra parameter."  When we added the second parameter, `tag`, we
said, "This might get out of hand, but there's no time to fix it now."

A third parameter is on the horizon, so let's take the opportunity to
add a "context" struct now to house all the additional transaction
data.