feat(generate): async streaming reads for `*Stub`
The generator can add asynchronous streaming reads in the `*Stub`
classes and their decorators.  No code is generated for the
`*Connection` or `*Client` classes.  We still don't know what is the
right API at that layer, and the services that need this have
hand-crafted `*Connection` and `*Client` layers at the moment.