fix(storage): avoid crashes when parsing `ErrorInfo`
Maybe unsurprisingly, the error payloads are less structured that we
expected. This PR performs additional validation to avoid crashes while
trying to parse these payloads as a `ErrorInfo`.  If we could reliably
use exceptions I would just parse the code and handle any exception in
a single line of code, alas, we cannot.