fix(GCS+gRPC): unused mutable Object fields (#8998)
A few fields in the `WithObjectMetadata()` option were not used. Notably
`custom_time()` and `acl()`.  I changed the unit test to make it easier
to inspect that all fields are tested, the expectation is created via
`ExpectedFullObjectMetadata()`.  A few fields are complicated, because
they only setable in requests that create new objects.

To fix the problem I refactored the code to copy mutable fields from the
`storage::*Request` classes to the `google.storage.v2.Object` proto.