feat(common): support environmental override for AuthorityOption
For each "GOOGLE_CLOUD_CPP_.*_ENDPOINT" environment variable used to
override `EndpointOption`, heed a new "GOOGLE_CLOUD_CPP_.*_AUTHORITY"
environment variable used to override `AuthorityOption`.

Both may need to be employed to communicate with an alternate backend.
This is true for the Spanner staging service, and for which we add a
breadcrumb to remind ourselves what needs to be tweaked to use it.