cleanup(GCS+gRPC): the environment variable overrides endpoint
Change the behavior default configuration for `EndpointOption` to
agree with the general principles for all libraries.  In this case,
this mean the environment variable overrides any settings via
`google::cloud::Options`.

The code was confusing because the existing environment variable
`CLOUD_STORAGE_GRPC_ENDPOINT` looks like a general purpose override,
separate from the `CLOUD_STORAGE_EMULATOR_ENDPOINT`. In practice, it was
only used to set the emulator endpoint. I simplified the code and
changed the environment variable (no notice required in experimental
components) to `CLOUD_STORAGE_TESTBENCH_GRPC_ENDPOINT`.

PS: is not possible to use a single environment variable for
both gRPC and REST, as the testbench uses different ports for each.
PPS: We generally prefer `TESTBENCH` over `EMULATOR` because the
latter conveys a level of fidelity we do not try to achieve.
Unfortunately we deprecated `CLOUD_STORAGE_TESTBENCH_ENDPOINT` and then
realized that "emulator" was a bad idea. We should fix that, but not in
this PR.