ci: enable GCS+gRPC production integration tests
The integration tests against production were disabled while the protos
stabilized. They are no anticipated breaking changes at this point.

We cannot pass the options created for a `GrpcClient` to a `CurlClient`.
The latter has a very subtle interpretation of `AuthorityOption` as it
supports two separate services: `iamcredentials.googleapis.com` is
used to sign blobs.

A number of tests remain disabled because they use metadata operations
in gRPC, which are not ready in production. Changed their TODO entries
to point to the right bug.