doc: update generated endpoint documentation during generate-libraries (#9056)
Teach the `generate-libraries` build how to update the generated endpoint
documentation.

Run it to update the `baremetalsolution` doc, which is the only one that
still has the required injection markers.  A followup PR will re-add these
markers to all the other `main.dox` files.

This, and the recent addition of `external/googleapis/update_libraries.sh`
to the `generate-libraries` build, means the the "How-to Guide" for adding
a generated library can be simplified, but that remains a TODO for now.