test(spanner): limit special instance-config filtering to prod
When selecting an `InstanceConfig` to use in the samples that create
a new `Instance`, we have some special filtering to avoid locations
where we run integration tests.  This filtering is specific to the
production Spanner service, and so should not be applied when using
an alternative like the staging service or the Spanner emulator.

This allows the `spanner_create_database_with_encryption_key` sample
to run on the staging service, as its encryption key is tied to a
location whose `InstanceConfig` was previously discarded by the filter.