feat(GCS+gRPC): timeout downloads on lack of progress (#9072)
Previously we timed out a download if it failed to complete the download
in a given time. This requires setting excessive timeouts for
applications that download both small objects (which take milliseconds)
and large objects (which can take minutes).

Furthermore, some applications start a download and then consume the
data slowly.  The timeout values for such applications would need to be
too large too.

This PR uses the asynchronous API to download objects. This API allows
us to timeout a `Read()` call, and abort the download if it is not
making progress.