cleanup: simplify quickstart CMake files
The CMake files for some libraries still hard-coded C++11, we do not
need to do this, as all the compilers we support (except Apple's Clang)
default to C++14 or higher. With Apple's Clang, Abseil requires C++11,
so that takes cares of the problem.

In addition, the quickstart projects do not require a "C" compiler, so
stop asking for one.