cleanup: modernize packaging instructions (#9099)
* cleanup: modernize packaging instructions

Our packaging instructions used `-H.` to set the source directory for
the CMake configuration step.  `-H` is an undocumented option, starting
with CMake 3.13 we can use `-S .`.

Likewise, our packaging instructions included `-DCMAKE_CXX_STANDARD=11`
for Abseil. This is not needed because the compilers we support all
default to C++14 or higher.

Finally, in some distros we compiled all libraries except gRPC as shared
libraries.  That seems unnecessarily inconsistent and confusing.

* Fix build for distros with old CMake versions

* Argh, regenerate markdown