feat: if available use target features to request C++ >= 11
As long as CMake >= 3.8, this will automatically require C++ >= 11 to
compile the library and their dependencies.  This is only relevant with
Apple Clang, the last compiler we support that defaults to C++98 (and
apparently will always default to C++98).  All other compilers we
support default to C++14 or higher.

While this only works for CMake >= 3.8, I expect this is almost always
the case on macOS.  In other words, this can fail to automate things in
theory, but should work well in practice.