fix(storage): do not duplicate debugging headers
The `storage::ObjectReadStream` class exposes debugging information via
the `std::multimap<std::string, std::string> headers()` function. Note
that this is a **multi** map because headers can be duplicated in some
cases.  At some point I made a mistake and returned an extra
`:curl-peer` header in each `ObjectReadSource::Read()` call.

If the application is downloading a very large object (say 100GiB) and
the application processes the data in small chunks (say a few KiB) then
millions of calls to `ObjectReadSource::Read()` would take place.
Consequently, the multimap contained millions of duplicate entries.

Previously this went unnoticed because (1) the multimap is released
when the download completes, and (2) our benchmarks and tests tend to
perform larger reads (say a few MiB) from objects that are at most a
few GiB in size. The memory did not grow enough to be noticeable.