feat(common): include status error payloads in the RPC log (#9169)
Include the `google.rpc.Status.details` messages when adding a failing
`Status` value to the RPC log.

See https://cloud.google.com/apis/design/errors#error_payloads for the
details on what is available.

`Status` already has smarts for dealing with a `google.rpc.ErrorInfo`
message.  I was never sure why this support didn't extend to the other
error-detail messages, but @devjgm assured me that `ErrorInfo` was all
that was required for user access.

But let's at least add all the available information to the RPC log.
This will help us understand failures, like the notoriously inscrutable
"Invalid request", where we'll now be able to see which field was bad
and a description of what went wrong.