feat(storage): per-operation options for `CurlClient`
Use the options from `CurrentOptions()` to control the behavior of
`CurlClient`.  In the implementation of some long-lived objects
(streaming writes, streaming reads), we need to copy the options used
when the object is created, and then set the span before making RPCs. I
tried to make this copy as soon as possible.

Some tests bypass `storage::Client` and need to be adjusted to create a
span. Ditto for some standalone functions, like parallel uploads. There
is an existing bug to add better support for per-operation options for
the standalone functions too.