impl(storage): use `CurrentOptions()` for `RetryClient`
The most interesting options in this case are the `RetryPolicyOption`
and `BackoffPolicyOption`.

Because now the policies are obtained from the `CurrentOptions()` and
these are initialized from `mock->options()`, the mocks need to return
sensible values for the options.

Some tests need to set the OptionsSpan, or the unit under test will not
work.