fix: relax parsing for GCE informational fields
We parse (but largely ignore) some metadata information about the GCE
service account. This parsing was very strict, rejecting the data if
fields were missing.  We now accept missing fields, and field them
with blank data (or empty lists).  We also accept the responses from
Cloud Shell, where the `scopes` field is a newline separated string,
instead of an array of strings.  This may be a bug in Cloud Shell, but
it makes the library unusable.