fix: handle transient failures in the GCE credentials
Getting the authorization token from the GCE credentials can fail
when the credentials are created.  The metadata service may be
experiencing some kind of transient failure.  It is better to return the
credentials, and let the retry loops handle the problem.  If the failure
is not transient, the retry loop will handle that too.