test(spanner): workaround for CreateBackup() timeout in samples
Apply a workaround to the samples, similar to that in the integration
tests, for the #8616 CreateBackup() timeout issue.  If we see the exact
error, we abandon the test but return a "success" status.

This does hide the problem from our view, but there is nothing the
client library can really do about it.  At least we can still see the
RPC trace should we look at the logs.

We're awaiting guidance on exactly how we should exercise operations
that may take hours to complete.