impl(GCS+gRPC): async accumulator for `ReadObject()` (#9277)
The asynchronous version of `ReadObject()` will return all the bytes. We
need a function to asynchronously accumulate all the values from a
single `AsyncReadObject()` call to the `StorageStub`. This function does
to retry or resume the download, that will come later. In case you are
wonder, we will require applications to only make "ranged reads", where
the number of bytes downloaded is controlled by the application.