ci(kokoro/macOS): use GCS for vcpkg binary cache
vcpkg can use GCS to keep its binary cache (i.e. pre-built binaries of
each package).  This cache uploads each package as it is built, which
is nice when doing a large rebuild, as the results are immediately
saved.  This is a content-based cache, and it includes the signature of
the compiler and other elements in the toolchain.