cleanup: no CMAKE_CXX_STANDARD in quickstarts
The generated quickstarts never set CMAKE_CXX_STANDARD, but we used to
set this in older code.  We also do not compile any "C" code so the
CMake projects do not need to waste time looking for a "C" compiler.

Also cleaned up one of the CI support files, which had a similar
set of problems.